chore: Move auth_util to shared_utils
