+++ data_resource/admin/routes/tableschema.py
@@ -3,7 +3,7 @@ from flask import Blueprint
 from data_resource.db import db_session
 import data_resource.admin.models as orm
 from data_resource.shared_utils import LogFactory
-from data_resource.admin.routes.auth_util import check_auth
+from data_resource.shared_utils.auth_util import check_auth
 
 
 tableschema_bp = Blueprint("tableschema_bp", __name__)
