+++ data_resource/shared_utils/auth_util.py
